Translatable literal (t_literal) should be casted to the primitive type defined in the field definition.
